requirements.txt
duckdb
numpy
pandas
joblib
mlflow
xgboost
scikit-learn
fastapi
uvicorn
pydantic
kafka-python
requests
pyarrow
evidently
prometheus-client
scipy
imbalanced-learn